requirements.txt
requests
python-dotenv
vext
vext.gi